Refactor clock/scanner/detector code
The 3 areas are now in separate files, so that the flow of setting up
each task is clearer. This should facilitate further decopuling the
clock, scanner, and detector in the future.

When settings changed, we keep track of what needs to be reconfigured
explicitly, rahter than flags that record the change itself. This
simplifies the management of reconfiguration.

An effort was made to minimize functional changes, but there may be a
few.