Decouple input reading from line size
Refactor the detector input processing, so that every time the input
callback is called, it reads all available samples in the buffer, not
just a single line's worth.

For now, the callback is still called after every line, though it should
be possible to change this wihtout any issue.

Also eliminate all the unnecessary intermediate buffers.